orchestration/requirements.txt
apache-airflow
apache-airflow-providers-docker
apache-airflow-providers-microsoft-azure